Restructured package into definition and inference subpackages
