+++ probeye/inference/__init__.py
@@ -1,0 +1,2 @@
+# subpackage imports
+from probeye.inference import taralli_
