--- probeye/solver/__init__.py
@@ -1,2 +1,0 @@
-# module imports
-from probeye.solver import taralli
